move server to cmd folder, version api folder
