+++ Dockerfile
@@ -4,7 +4,7 @@ WORKDIR /go/src/github.com/dairycart/dairycart
 
 ADD . .
 
-RUN go build -o /dairycart
+RUN go build -o /dairycart github.com/dairycart/dairycart/cmd/server/v1
 
 # final stage
 FROM ubuntu:latest
