+++ cmd/integration_tests/v1/Dockerfile
@@ -1,7 +1,8 @@
 FROM golang:latest
 WORKDIR /go/src/github.com/dairycart/dairycart
 
-ADD integration_tests .
+# this is meant to be run in the upper context
+ADD cmd/integration_tests/v1 .
 ADD vendor vendor
 
 ENTRYPOINT ["go", "test", "-v"]
