+++ cmd/integration_tests/v1/auth_test.go
