+++ cmd/integration_tests/v1/manifest.json
