+++ integration-tests.yml
@@ -34,7 +34,7 @@ services:
   test:
     build:
       context: .
-      dockerfile: integration_tests/Dockerfile
+      dockerfile: cmd/integration_tests/v1/Dockerfile
     links:
       - api
       - httpbin
