+++ test.Dockerfile
@@ -3,4 +3,4 @@ WORKDIR /go/src/github.com/dairycart/dairycart
 
 ADD . .
 
-ENTRYPOINT ["go", "test", "-cover", "github.com/dairycart/dairycart/api", "-parallel=1"]
+ENTRYPOINT ["go", "test", "-cover", "github.com/dairycart/dairycart/api/v1", "-parallel=1"]
